Surfaces can now build lists of their neighboring cells. The geometry class can now call a function which will compute the cells which neighbor surfaces on their positive or negative sides. This should help us speed up the track segmentation process since we will want to quickly find which cell a track enters next after it crosses a surface. This method will only work after the geometry is parsed and the Geometry's adjustKeys is called

+++ src/Lattice.cpp
@@ -123,8 +123,8 @@ void Lattice::adjustKeys(std::map<int, Universe*> universes) {
 	std::vector< std::vector<int> > adjusted_universes;
 
 	/* Adjust the indices for each universe to be the universe's uid */
-	for (int U = 0; U < _universes.size(); U++) {
-		for (int u = 0; u < _universes.at(U).size(); u++) {
+	for (int U = 0; U < (int)_universes.size(); U++) {
+		for (int u = 0; u < (int)_universes.at(U).size(); u++) {
 			int universe = _universes.at(U).at(u);
 			adjusted_universes[U][u] = universes.at(universe)->getUid();
 		}
